Completed making only one name appear in file list at a time. Able to grab the log file base name. Unable to get PHP download action to work. Throws nullPointer exception on converting InputStream to BufferedReader. Taking a break. Latest commit on github.
Thinking that may have to change the HttpRequest type to GET and handle
the download through a GET Request. Hopefully That will fix the problem.

Have implemented Exception catches for NullPointer, Interrupted (due to
wait for stream to become available) in NetworkSetupActivity.

The NullPointerExecption is being thrown at line 129 of
NetworkSetupActivity.java.